structure reformart
